Saving change for push to github

+++ pkg/auth/controller/constants.go
@@ -1,5 +1,5 @@
 package controller
 
 var (
-	SESSION_EXPIRE_IN_SECOND int64 = 3600
+	SESSION_EXPIRE_IN_SECOND int = 3600
 )
